+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All entries in reverse chronological order.
 Version 0.9 (2020-05-??)
 
 - feat, l4: distinguish p3 slopes fig from p4 chi-squared plots and merge together
+- enh, l4: chisquared detail plots to have region
 
 
 Version 0.8 (2020-05-13)